코드 최적화+정리
